# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paper Company",
    "version": "1.0.0",
    "description": "Witaj w repozytorium Systemu ERP dla Papier Company - zainspirowanego przez kultowy serial \"The Office\". Nasz system jest zaprojektowany, aby wspierać i optymalizować kluczowe procesy biznesowe w firmie zajmującej się dystrybucją papieru, od zarządzania zapasami po finanse i księgowość."
  },
  "paths": {
...
    "/auth": {
      "post": {
        "summary": "Logowanie użytkownika",
        "description": "Loguje użytkownika i zwraca token dostępowy.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Nazwa użytkownika."
                  },
                  "password": {
                    "type": "string",
                    "description": "Hasło użytkownika."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomyślnie zalogowano użytkownika.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Token dostępowy."
                    },
                    "idUser": {
                      "type": "string",
                      "description": "Identyfikator użytkownika."
                    },
                    "message": {
                      "type": "string",
                      "description": "Komunikat powitalny."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Błędne żądanie.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Błędna nazwa użytkownika lub hasło.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Konto nieaktywne lub zablokowane.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Nieprawidłowe dane wejściowe.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Odśwież token dostępowy",
        "description": "Odświeża token dostępowy użytkownika na podstawie tokena odświeżającego.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idUser": {
                    "type": "string",
                    "description": "Identyfikator użytkownika."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomyślnie odświeżono token dostępowy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Nowy token dostępowy."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Brak tokena odświeżającego lub błąd weryfikacji.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Nieprawidłowy token odświeżający.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Pobiera listę wszystkich użytkowników.",
        "description": "Endpoint dostępny tylko dla użytkowników o roli \"admin\".",
        "tags": [
          "nn"
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano listę użytkowników.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usersList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": "// Odpowiednie pola, takie jak id, username, email itp."
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Pobiera informacje o użytkowniku.",
        "description": "Endpoint dostępny tylko dla użytkowników o roli \"admin\".",
        "tags": [
          "nn"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano informacje o użytkowniku.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": "// Odpowiednie pola, takie jak id, username, email itp."
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/user/{id}": {
      "get": {
        "summary": "Pobiera informacje o użytkowniku.",
        "description": "Pobiera informacje o użytkowniku na podstawie jego identyfikatora.",
        "tags": [
          "nn"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano informacje o użytkowniku.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}